June-2024

+++ 07 - Algorithms & Problem Solving Level 3/04 _ Sum Each Column In Matrix/04 _ Sum Each Column In Matrix/04 _ Sum Each Column In Matrix.cpp
@@ -6,13 +6,13 @@ int GenerateRandomNumber(int from, int to) {
 	return rand() % (to - from + 1) + from;
 }
 
-void FillMatrixWithRandomNumber(int arr[3][3], int rows, int cols) {
+void FillMatrixWithRandomNumber(int arr[3][3], short rows, short cols) {
 	for (int i = 0; i < rows; i++)
 		for (int j = 0; j < cols; j++)
 			arr[i][j] = GenerateRandomNumber(1, 100);
 }
 
-void PrintMatrix(int arr[3][3], int rows, int cols) {
+void PrintMatrix(int arr[3][3], short rows, short cols) {
 	for (int i = 0; i < rows; i++) {
 		for (int j = 0; j < cols; j++)
 			cout << arr[i][j] << " ";
@@ -20,14 +20,14 @@ void PrintMatrix(int arr[3][3], int rows, int cols) {
 	}
 }
 
-int ColumnsSum(int arr[3][3], int rows, int cols) {
+int ColumnsSum(int arr[3][3], short rows, short cols) {
 	int sum = 0;
 	for (int i = 0; i < rows; i++)
 		sum += arr[i][cols];
 	return sum;
 }
 
-void PrintEachColumnsSum(int arr[3][3], int rows, int cols) {
+void PrintEachColumnsSum(int arr[3][3], short rows, short cols) {
 	for (int i = 0; i < cols; i++)
 		cout << "column " << i + 1 << " = " << ColumnsSum(arr, rows, i) << endl;
 }
